Collider2D
AABB알고리즘을 사용해서 감지합니다
원본 => m_AABB, 대상 => d_AABB
//d_AABB의 가장왼쪽이 x의 최대 범위안에 들어오고 m_AABB도 이와 같다면 ==> x축은 겹침
이런식으로 x축과 y축이 모두 겹치면 true를 반환합니다
매번 감지하면 비효율적일것 같아서 일부 유니티 기능을 본따 버튼을 누를때만 충돌을 감지하도록 만들고 있습니다
지금은 충돌체의 값을 넘겨주는 기능을 만들고 있습니다

+++ include/Vector2D.h
@@ -19,7 +19,7 @@ public:
 	//pow(double base, double n) ==> base가 되는 숫자의 n 제곱, 즉 제곱을 구하는 함수 
 	//pow(5, 3) -> 10의 2제곱 
 	//둘모두 함수 오버로딩에 의해 double외의 다른 자료형 사용 가능
-	float length() { return sqrt((float)(m_x * m_x + m_y * m_y)); } //둘다 10, 10이면 200 = 16.xxxxx
+	float length() { return sqrt(m_x * m_x + m_y * m_y); } //둘다 10, 10이면 200 = 16.xxxxx
 
 	Vector2D operator+(const Vector2D& v2) const;
 	Vector2D operator-(const Vector2D& v2) const;
